feat(tooling): enforce commitlint (Closes #50)
Take example on [this issue](https://gitlab.com/gitlab-org/release-cli/-/issues/112) to add commitlint in the project pipelines.

The pipeline is designed to work on merge_request pipeline as merge_result pipelines is a paid feature of Gitlab.

The pipeline tends to replicate the result of the project squash merge template to verify that it follows conventional commit rules

Close #50